RavenDB-19602 fixed paging in RawQuery
